feat(types): More ssz conversion (#1784)

+++ mod/engine-primitives/pkg/engine-primitives/withdrawal_ssz_test.go
@@ -44,7 +44,7 @@ func TestWithdrawalSSZ(t *testing.T) {
 	require.NoError(t, err)
 
 	size := withdrawal.SizeSSZ()
-	require.Equal(t, 44, size)
+	require.Equal(t, uint32(44), size)
 
 	tree, errHashTree := withdrawal.HashTreeRoot()
 	require.NoError(t, errHashTree)
